// Project/gold-loan/src/app/model/customerdetail.model.ts
export class customerdetail{
    customercodepopup: number;
    firstname: string;
    lastname: string;
    locality: string;
    address1: string;
    
    address2: string;
    district: string;
    state: string;
    pin: string;
    dateofbirth: Date;
    gender: string;
    mobilenumber: number;
    landlinenumber: number;
    email: string;
    photoPath?: string;
}